+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - `CONMAN_ADMIN_ROUTES` setting has been removed. In future, we'll
   automatically detect subclasses of `Route` for admin integration.
+- Renamed `RouteViewHandler` to `ViewHandler`.
 
 ### Changed